# archive/v1_2025/api/openapi.yaml
openapi: 3.0.3
info:
  title: ELWOSA API
  description: |
    Enterprise-Level Work Organization & System Architecture API
    
    Diese API bietet umfassende Funktionen für Projektmanagement,
    Task-Verwaltung und KI-gestützte Workflows.
    
    ## Authentifizierung
    
    Die API verwendet JWT-basierte Authentifizierung. Alle Endpunkte
    (außer Login/Register) erfordern einen gültigen Bearer Token.
    
    ## Rate Limiting
    
    - 1000 Requests pro Stunde für authentifizierte Benutzer
    - 100 Requests pro Stunde für nicht-authentifizierte Benutzer
    
    ## Fehlerbehandlung
    
    Die API gibt standardisierte HTTP-Statuscodes zurück:
    - 200: Erfolg
    - 400: Bad Request (Validierungsfehler)
    - 401: Unauthorized (Authentifizierung erforderlich)
    - 403: Forbidden (Unzureichende Berechtigung)
    - 404: Not Found
    - 429: Too Many Requests (Rate Limit erreicht)
    - 500: Internal Server Error
    
  version: 1.0.0
  contact:
    name: ELWOSA Support
    url: https://github.com/MadGapun/ELWOSA-Pub
    email: support@elwosa.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8001
    description: Lokale Entwicklung
  - url: https://api.elwosa.com/v1
    description: Produktions-API
security:
  - BearerAuth: []
paths:
...
  /tasks:
    get:
      tags:
        - Tasks
      summary: Tasks abrufen
      description: Gibt eine Liste aller Tasks des authentifizierten Benutzers zurück
      parameters:
        - name: status
          in: query
          description: Filtert Tasks nach Status
          schema:
            type: string
            enum: [open, in_progress, completed, cancelled]
        - name: priority
          in: query
          description: Filtert Tasks nach Priorität
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: assigned_to
          in: query
          description: Filtert Tasks nach zugewiesenem Benutzer
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximale Anzahl zurückgegebener Tasks
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Anzahl zu überspringender Tasks für Paginierung
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Liste der Tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                    description: Gesamtanzahl der Tasks
                  limit:
                    type: integer
                  offset:
                    type: integer
    
    post:
      tags:
        - Tasks
      summary: Neue Task erstellen
      description: Erstellt eine neue Task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task erfolgreich erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Validierungsfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{task_id}:
    get:
      tags:
        - Tasks
      summary: Task Details abrufen
      description: Gibt Details einer spezifischen Task zurück
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID der Task
          schema:
            type: integer
      responses:
        '200':
          description: Task Details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task nicht gefunden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Tasks
      summary: Task aktualisieren
      description: Aktualisiert eine bestehende Task
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID der Task
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task nicht gefunden
    
    delete:
      tags:
        - Tasks
      summary: Task löschen
      description: Löscht eine Task
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID der Task
          schema:
            type: integer
      responses:
        '204':
          description: Task erfolgreich gelöscht
        '404':
          description: Task nicht gefunden
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 123
        title:
          type: string
          example: "API Dokumentation aktualisieren"
        description:
          type: string
          example: "OpenAPI 3.0 Spezifikation für neue Endpunkte erstellen"
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
          example: in_progress
        priority:
          type: integer
          minimum: 1
          maximum: 5
          example: 2
        assigned_to:
          type: integer
          example: 42
        project_id:
          type: integer
          example: 7
        estimated_hours:
          type: number
          example: 4.5
        actual_hours:
          type: number
          example: 3.2
        created_at:
          type: string
          format: date-time
          example: "2025-01-15T09:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-15T14:30:00Z"
        completed_at:
          type: string
          format: date-time
          nullable: true
          example: null
        steps:
          type: array
          items:
            $ref: '#/components/schemas/TaskStep'
    TaskCreate:
      type: object
      required:
        - title
        - project_id
      properties:
        title:
          type: string
          maxLength: 255
          example: "Neue Feature implementieren"
        description:
          type: string
          example: "Detaillierte Beschreibung der Task"
        priority:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
        assigned_to:
          type: integer
          example: 42
        project_id:
          type: integer
          example: 7
        estimated_hours:
          type: number
          minimum: 0
          example: 8.0
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        priority:
          type: integer
          minimum: 1
          maximum: 5
        assigned_to:
          type: integer
        estimated_hours:
          type: number
          minimum: 0
        actual_hours:
          type: number
          minimum: 0
    TaskStep:
      type: object
      properties:
        id:
          type: integer
          example: 456
        task_id:
          type: integer
          example: 123
        content:
          type: string
          example: "Endpoint /users implementiert und getestet"
        user:
          type: string
          example: "Papa Claude"
        timestamp:
          type: string
          format: date-time
          example: "2025-01-15T11:45:30Z"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Das Feld 'title' ist erforderlich"
        code:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time
          example: "2025-01-15T14:30:00Z"
        details:
          type: object
          additionalProperties: true
          example:
            field: "title"
            constraint: "required"
...